Refactor comments popup function to work after the items are loaded
